Refactor key*gen: share data structure
This simply moves 'struct register_internal' into 'keygen/keygen_data.h'
